diplom/source/src/requirements.txt
Django==1.6
Jinja2==2.7.1
MarkupSafe==0.18
NoseGAE==0.2.0
PIL==1.1.7
PyBrain==0.3
PyNN==0.7.5
Pygments==1.6
Sphinx==1.2
WebOb==1.3.1
WebTest==2.0.11
argparse==1.2.1
astroid==1.0.1
beautifulsoup4==4.3.2
docutils==0.11
google-api-python-client==1.2
httplib2==0.8
logilab-common==0.60.1
networkx==1.8.1
nose==1.3.0
numpy==1.6.1
pandas==0.12.0
ply==3.4
pylint==1.1.0
pystache==0.5.3
python-dateutil==2.2
pytz==2013.8
six==1.4.1
sphinx-bootstrap-theme==0.3.4
waitress==0.8.8
webapp2==2.5.2
wsgiref==0.1.2
